requirements.txt
beautifulsoup4~=4.12.2
selenium~=4.8.3
twilio~=8.0.0
python-dotenv~=1.0.0
webdriver-manager~=3.8.6
bs4~=0.0.1